Add fromThrowable utility function
It is common within javascript libraries to throw within a synchronous function
This adds a utility function fromThowable that wraps a function and creates a
function with the same signature except instead of throwing, a Result will be
returned. Ok if the original function succeds, or Err if the function throws.

As we cannot infer the type of the thrown error, this will be of unknown type
unless a mapping function is provided.